Port kot-ui++

+++ Sources/Libs/mlibc/Src/sysdeps/kot/include/kot/uisd/srvs/audio.h
@@ -88,7 +88,7 @@ typedef struct{
 
 struct kot_srv_audio_callback_t{
     kot_thread_t Self;
-    uint64_t Data;
+    uintptr_t Data;
     size64_t Size;
     bool IsAwait;
     KResult Status;
