+++ Sources/Libs/mlibc/Src/sysdeps/kot/include/kot/uisd/srvs/time.h
@@ -26,7 +26,7 @@ typedef struct{
 
 struct kot_srv_time_callback_t{
     kot_thread_t Self;
-    uint64_t Data;
+    uintptr_t Data;
     size64_t Size;
     bool IsAwait;
     KResult Status;
